chore: pr comments addressed
